requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
dropbox==12.0.2
pdfplumber==0.11.4
PyMuPDF==1.24.14
python-docx==1.1.2
python-dateutil==2.9.0.post0
email-validator==2.2.0
